Avoid 'conversion may change value' cases related to JobMan
Avoid two warnings related to JobMan confirming the conversions to `int`
because:
- A non-Krusader function that is used (`setValue(int value)`) only accepts
that type.
- An existing `int percent()` function only returns an `int` (and it doesn't
need returning numbers bigger than 100).

Revision: https://invent.kde.org/utilities/krusader/-/merge_requests/57

+++ krusader/JobMan/jobman.cpp
@@ -86,7 +86,7 @@ protected slots:
         }
     }
 
-    void slotPercent(KJob *, unsigned long percent) { m_progressBar->setValue(percent); }
+    void slotPercent(KJob *, unsigned long percent) { m_progressBar->setValue(static_cast<int>(percent)); }
 
     void updatePauseResumeButton()
     {
